+++ test/grax/schema/inheritance_test.exs
@@ -1,7 +1,7 @@
 defmodule Grax.Schema.InheritanceTest do
   use Grax.TestCase
 
-  alias Grax.Schema.Inheritance
+  alias Grax.Schema.{Inheritance, TypeError}
   alias Grax.InvalidResourceTypeError
 
   alias Example.{
@@ -11,7 +11,8 @@ defmodule Grax.Schema.InheritanceTest do
     ChildSchema,
     ChildSchemaWithClass,
     ChildOfMany,
-    LinksWithInheritedSchemas
+    LinksWithInheritedSchemas,
+    NonPolymorphicLink
   }
 
   test "__super__/0" do
@@ -73,6 +74,7 @@ defmodule Grax.Schema.InheritanceTest do
                  iri: ~I<http://example.com/lp1>,
                  schema: ChildSchema,
                  on_type_mismatch: :ignore,
+                 polymorphic: true,
                  type: {:resource, User}
                },
                lp2: %Grax.Schema.LinkProperty{
@@ -80,6 +82,7 @@ defmodule Grax.Schema.InheritanceTest do
                  iri: ~I<http://example.com/lp22>,
                  schema: ChildSchema,
                  on_type_mismatch: :ignore,
+                 polymorphic: true,
                  type: {:resource, User}
                },
                lp3: %Grax.Schema.LinkProperty{
@@ -87,6 +90,7 @@ defmodule Grax.Schema.InheritanceTest do
                  iri: ~I<http://example.com/lp3>,
                  schema: ChildSchema,
                  on_type_mismatch: :ignore,
+                 polymorphic: true,
                  type: {:resource, User}
                }
              }
@@ -136,6 +140,7 @@ defmodule Grax.Schema.InheritanceTest do
                  iri: ~I<http://example.com/lp1>,
                  schema: ChildOfMany,
                  on_type_mismatch: :ignore,
+                 polymorphic: true,
                  type: {:resource, User}
                },
                lp2: %Grax.Schema.LinkProperty{
@@ -143,6 +148,7 @@ defmodule Grax.Schema.InheritanceTest do
                  iri: ~I<http://example.com/lp2>,
                  schema: ChildOfMany,
                  on_type_mismatch: :ignore,
+                 polymorphic: true,
                  type: {:resource, User}
                },
                lp3: %Grax.Schema.LinkProperty{
@@ -150,6 +156,7 @@ defmodule Grax.Schema.InheritanceTest do
                  iri: ~I<http://example.com/lp3>,
                  schema: ChildOfMany,
                  on_type_mismatch: :ignore,
+                 polymorphic: true,
                  type: {:resource, User}
                },
                lp4: %Grax.Schema.LinkProperty{
@@ -157,6 +164,7 @@ defmodule Grax.Schema.InheritanceTest do
                  iri: ~I<http://example.com/lp4>,
                  schema: ChildOfMany,
                  on_type_mismatch: :ignore,
+                 polymorphic: true,
                  type: {:resource, User}
                }
              }
@@ -228,6 +236,22 @@ defmodule Grax.Schema.InheritanceTest do
                    ChildOfMany.build!(EX.C)
                  ]
                )
+    end
+
+    test "non-polymorphic property with matching schema" do
+      assert NonPolymorphicLink.build!(EX.A)
+             |> Grax.put(:link, ParentSchema.build!(EX.B)) ==
+               NonPolymorphicLink.build(EX.A, link: ParentSchema.build!(EX.B))
+    end
+
+    test "non-polymorphic property with inherited schema" do
+      assert NonPolymorphicLink.build!(EX.A)
+             |> Grax.put(:link, ChildSchemaWithClass.build!(EX.B)) ==
+               {:error,
+                TypeError.exception(
+                  value: ChildSchemaWithClass.build!(EX.B),
+                  type: ParentSchema
+                )}
     end
   end
 
@@ -342,6 +366,61 @@ defmodule Grax.Schema.InheritanceTest do
                 InvalidResourceTypeError.exception(
                   type: :no_match,
                   resource_types: [RDF.iri(EX.Unknown)]
+                )}
+    end
+
+    test "non-polymorphic property with inherited and parent schema" do
+      assert RDF.graph([
+               EX.A |> EX.link(EX.B),
+               EX.B |> RDF.type([EX.Parent, EX.Child2])
+             ])
+             |> NonPolymorphicLink.load(EX.A) ==
+               NonPolymorphicLink.build(EX.A,
+                 link:
+                   ParentSchema.build!(EX.B,
+                     __additional_statements__: %{RDF.type() => [EX.Parent, EX.Child2]}
+                   )
+               )
+    end
+
+    test "non-polymorphic and non-strict property with inherited schema only" do
+      assert RDF.graph([
+               EX.A |> EX.strictLink(EX.B),
+               EX.B |> RDF.type(EX.Child2)
+             ])
+             |> NonPolymorphicLink.load(EX.A) ==
+               NonPolymorphicLink.build(EX.A,
+                 strict_link:
+                   ParentSchema.build!(EX.B,
+                     __additional_statements__: %{RDF.type() => [EX.Parent, EX.Child2]}
+                   )
+               )
+    end
+
+    test "non-polymorphic strict property with inherited schema only" do
+      assert RDF.graph([
+               EX.A |> EX.strictLink(EX.B),
+               EX.B |> RDF.type(EX.Child2)
+             ])
+             |> NonPolymorphicLink.load(EX.A) ==
+               NonPolymorphicLink.build(EX.A,
+                 strict_link:
+                   ParentSchema.build!(EX.B,
+                     __additional_statements__: %{RDF.type() => [EX.Parent, EX.Child2]}
+                   )
+               )
+    end
+
+    test "non-polymorphic strict property with none matching schema" do
+      assert RDF.graph([
+               EX.A |> EX.strictLink(EX.B),
+               EX.B |> RDF.type(EX.User)
+             ])
+             |> NonPolymorphicLink.load(EX.A) ==
+               {:error,
+                InvalidResourceTypeError.exception(
+                  type: :no_match,
+                  resource_types: [RDF.iri(EX.User)]
                 )}
     end
   end